requirements.txt
addict
fairscale
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch==2.5.0
torchvision
tqdm
yapf
